Restart 3dfier if tile fails
When the swap fill up the 3dfier process
cannot allocate memory and this process
fails. Some AHN tiles contain more points
than the others, thus has a higher memory
usage. This commit implements that the
failed tiles are restarted max 3 times. The
restarts are done when all the rest of the
tiles are processed, thus there will be more memory available for the
restarted tiles.